Fix #89 slow rendering due to transparency
Previously we were apply transparency even if there was no transparency in the pixels, this has been rectified. A method that approximates transparency has also been added for when drawing an ink with one layer that has transparency (a good example of this would be a PNG with transparency). This fake transparency is very fast and very useful for working with single layer transparent images.