Fix issue #86 Pygame bezier curves
This now replicates nifty.ink and draws the same bezier curves in Pygame to improve accuracy! There appears to be an antialiasing problem if you increase the stepsize too much, so I will work on implementing AA next! It's possible we can replace the generic bezier curve function with a quadratic one inside, but it may be useful in actually using fancy curves in inks themselves!