Major Pygame Improvements (#107)
* Fix Transparency Speed

* Ability to simulate drawing

* Fix #93 colour bug with point and square images

Issue was due to how PIL internally stored and processed colour values for different image formats.

* Fix #89 slow rendering due to transparency

Previously we were apply transparency even if there was no transparency in the pixels, this has been rectified. A method that approximates transparency has also been added for when drawing an ink with one layer that has transparency (a good example of this would be a PNG with transparency). This fake transparency is very fast and very useful for working with single layer transparent images.

* Fix #94 Pygame line thickness inconsistency

All diagonal lines previously had an incorrect width due to quirks of how lines are drawn in Pygame. This has been solved by using vectors to calculate the 4 points of a line with non-zero thickness

* Fix issue #86 Pygame bezier curves

This now replicates nifty.ink and draws the same bezier curves in Pygame to improve accuracy! There appears to be an antialiasing problem if you increase the stepsize too much, so I will work on implementing AA next! It's possible we can replace the generic bezier curve function with a quadratic one inside, but it may be useful in actually using fancy curves in inks themselves!

* Improve transparency speed slightly

Also sourced the bezier curve function and renamed some parameters and gave different default values.