Most PEP 8 Violations resolved
I avoided most violations relating to going over a line width of 120 characters, they should be handled where possible however.

We also have a lot of variables that shadow variables in outer scopes, this should be resolved in some cases, but general cases like 'drawing' should be fine to keep.