General Improvements (#108)
* Rename line functions

* Rename import functions

* Most PEP 8 Violations resolved

I avoided most violations relating to going over a line width of 120 characters, they should be handled where possible however.

We also have a lot of variables that shadow variables in outer scopes, this should be resolved in some cases, but general cases like 'drawing' should be fine to keep.

* Make size meaningful

Size now displays in megabytes instead of bytes.

* Add star example and comments

A drawing example I previously made but never included until now.

* Remove unnecessary imports

* Ability to round floats in a drawing

Ability to reduce the size of a drawing by rounding all the floats in the drawing data.

* Add random seed methods

A random seed can be displayed and changed now to make it possible to reproduce random drawings.

* Save and Load Drawings from files

Stored as JSON

* Add layer and add layer from file

Useful layer related methods. add_layer_from_file uses the exported raw data.

* Fix black transparency edge case

Black was inadvertently being removed due to the color key.

* Fix #49 More options for placing, scaling and rotating images

This includes point images and square images.

* Add general bezier curve line function

This function can be used to add fancy curves that nifty.ink and Pygame can't normally handle, such as cubic or quartic bezier curves.

* Fix #96 Add rectangle method, rectangle example and update square

This creates a new shape drawing method, it draws in the long direction of the rectangle and the square method was updated to use the rectangle method.

* Update square image to use rectangles

This fixed a slight width bug when drawing stretched images too.