Convert the repository to a python package
The repository is now a build environment for testing and building pyautonifty.
Only code inside the pyautonifty directory is included in the package.
main.py is simply a file to interact with the package now.

Examples have not been tested due to the huge overhaul to the code, however, they will be rewritten in the next commit which will create an example folder with standalone examples (mostly compiled from drawing_methods.py)