Remove and test examples from the package
All the drawing examples in drawing_methods.py have been moved into standalone python scripts that can execute and function as expected. This is much more in line with typical examples for a package (we don't want to waste space on examples in the actual package).

One example has some minor issues (squared_circle.py) but this will need to be fixed by Markichu or someone else with a keen eye :)